delivery working without update
